Pull out StdLib generation
