+++ libs/language-server/src/lib/builtin-library/stdlib.ts
@@ -1,0 +1,5 @@
+import { BuiltinValuetypesLib } from '../ast/wrappers/value-type/primitive/primitive-valuetypes';
+
+import { StdLib as PartialStdLib } from './generated/stdlib';
+
+export const StdLib = { ...PartialStdLib, ...BuiltinValuetypesLib };
